@@ -2,32 +2,28 @@ package br.uniesp.si.techback.controller;
 
 import br.uniesp.si.techback.dto.conteudo.ConteudoCreateDTO;
 import br.uniesp.si.techback.dto.conteudo.ConteudoResponseDTO;
-import br.uniesp.si.techback.dto.conteudo.ConteudoUpdateDTO;
+import br.uniesp.si.techback.model.enums.TipoConteudo;
 import br.uniesp.si.techback.service.ConteudoService;
-import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/conteudos")
+@RequestMapping("/api/v1/conteudos")
+@RequiredArgsConstructor
 public class ConteudoController {
 
     private final ConteudoService service;
 
-    public ConteudoController(ConteudoService service) {
-        this.service = service;
-    }
-
-    @PostMapping
-    public ConteudoResponseDTO criar(@Valid @RequestBody ConteudoCreateDTO dto) {
-        return service.criar(dto);
-    }
-
     @GetMapping
-    public List<ConteudoResponseDTO> listar() {
-        return service.listar();
+    public List<ConteudoResponseDTO> listar(
+            @RequestParam(required = false) TipoConteudo tipo,
+            @RequestParam(required = false) String genero,
+            @RequestParam(required = false) String q
+    ) {
+        return service.listar(tipo, genero, q);
     }
 
     @GetMapping("/{id}")
@@ -35,16 +31,21 @@ public class ConteudoController {
         return service.buscar(id);
     }
 
+    @PostMapping
+    public ConteudoResponseDTO criar(@RequestBody ConteudoCreateDTO dto) {
+        return service.criar(dto);
+    }
+
     @PutMapping("/{id}")
     public ConteudoResponseDTO atualizar(
             @PathVariable UUID id,
-            @Valid @RequestBody ConteudoUpdateDTO dto
+            @RequestBody ConteudoCreateDTO dto
     ) {
         return service.atualizar(id, dto);
     }
 
     @DeleteMapping("/{id}")
-    public void deletar(@PathVariable UUID id) {
-        service.deletar(id);
+    public void excluir(@PathVariable UUID id) {
+        service.excluir(id);
     }
 }
