@@ -1,15 +1,15 @@
 package br.uniesp.si.techback.dto.assinatura;
 
+import br.uniesp.si.techback.model.enums.StatusAssinatura;
+
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record AssinaturaResponseDTO(
         UUID id,
         UUID usuarioId,
-        UUID buscar,
         UUID planoId,
-        UUID metodoPagamentoId,
-        String status,
+        StatusAssinatura status,
         LocalDateTime iniciadaEm,
         LocalDateTime canceladaEm
 ) {}
