@@ -1,14 +1,16 @@
 package br.uniesp.si.techback.dto.conteudo;
 
-import jakarta.validation.constraints.*;
+import br.uniesp.si.techback.model.enums.TipoConteudo;
+
+import java.math.BigDecimal;
 
 public record ConteudoCreateDTO(
-        @NotBlank @Size(max = 200) String titulo,
-        @NotBlank @Pattern(regexp = "FILME|SERIE") String tipo,
-        @NotNull @Min(1888) @Max(2100) Integer ano,
-        @NotNull @Min(1) @Max(999) Integer duracaoMinutos,
-        @NotNull @DecimalMin("0.00") @DecimalMax("10.00") Double relevancia,
+        String titulo,
+        TipoConteudo tipo,
+        Integer ano,
+        Integer duracaoMinutos,
+        BigDecimal relevancia,
         String sinopse,
-        @Size(max = 500) String trailerUrl,
-        @Size(max = 50) String genero
+        String genero,
+        String trailerUrl
 ) {}
