@@ -1,14 +1,14 @@
 package br.uniesp.si.techback.dto.endereco;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
+import java.util.UUID;
 
 public record EnderecoCreateDTO(
-        @NotBlank @Size(max = 150) String logradouro,
-        @NotBlank @Size(max = 20) String numero,
-        @Size(max = 150) String complemento,
-        @NotBlank @Size(max = 100) String bairro,
-        @NotBlank @Size(max = 100) String cidade,
-        @NotBlank @Size(min = 2, max = 2) String estado,
-        @NotBlank @Size(min = 8, max = 9) String cep
+        UUID usuarioId,
+        String logradouro,
+        String numero,
+        String complemento,
+        String bairro,
+        String cidade,
+        String estado,
+        String cep
 ) {}
