@@ -5,11 +5,8 @@ import java.util.UUID;
 public record MetodoPagamentoResponseDTO(
         UUID id,
         UUID usuarioId,
-        String tipo,
-        String apelido,
         String bandeira,
-        String numeroMascarado,
-        String expiracao,
-        String chavePix,
-        boolean ativo
+        String ultimos4,
+        Integer mesExp,
+        Integer anoExp
 ) {}
