@@ -1,15 +1,13 @@
 package br.uniesp.si.techback.dto.usuario;
 
-import br.uniesp.si.techback.dto.endereco.EnderecoResponseDTO;
-import java.time.LocalDate;
+import br.uniesp.si.techback.model.enums.PerfilUsuario;
+
 import java.util.UUID;
 
 public record UsuarioResponseDTO(
         UUID id,
         String nomeCompleto,
-        LocalDate dataNascimento,
         String email,
-        String cpfCnpj,
-        String perfil,
-        EnderecoResponseDTO endereco
+        PerfilUsuario perfil,
+        String cpfCnpj
 ) {}
