@@ -1,21 +1,18 @@
 package br.uniesp.si.techback.model;
 
+import br.uniesp.si.techback.model.enums.TipoConteudo;
 import jakarta.persistence.*;
 import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(of = "id")
 @Entity
 @Table(name = "conteudo")
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor
+@EqualsAndHashCode(of = "id")
 public class Conteudo {
 
     @Id
@@ -25,8 +22,9 @@ public class Conteudo {
     @Column(nullable = false, length = 200)
     private String titulo;
 
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 10)
-    private String tipo;
+    private TipoConteudo tipo;
 
     @Column(nullable = false)
     private Integer ano;
@@ -34,23 +32,31 @@ public class Conteudo {
     @Column(name = "duracao_minutos", nullable = false)
     private Integer duracaoMinutos;
 
-    @Column(nullable = false, precision = 6, scale = 2)
-    private Double relevancia;
+    @Column(nullable = false, precision = 4, scale = 2)
+    private BigDecimal relevancia;
 
-    @Lob
+    @Column(columnDefinition = "TEXT")
     private String sinopse;
-
-    @Column(name = "trailer_url", length = 500)
-    private String trailerUrl;
 
     @Column(length = 50)
     private String genero;
 
-    @CreationTimestamp
-    @Column(name = "criado_em", nullable = false, updatable = false)
+    @Column(length = 500)
+    private String trailerUrl;
+
+    @Column(nullable = false)
     private LocalDateTime criadoEm;
 
-    @UpdateTimestamp
-    @Column(name = "atualizado_em", nullable = false)
+    @Column(nullable = false)
     private LocalDateTime atualizadoEm;
+
+    @PrePersist
+    void prePersist() {
+        criadoEm = atualizadoEm = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    void preUpdate() {
+        atualizadoEm = LocalDateTime.now();
+    }
 }
