@@ -5,17 +5,30 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
 @Entity
-@Table(name = "favoritos")
+@Table(name = "favorito")
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor
 public class Favorito {
 
     @EmbeddedId
     private FavoritoId id;
 
-    @Column(name = "criado_em", nullable = false)
-    private LocalDateTime criadoEm = LocalDateTime.now();
+    @ManyToOne
+    @MapsId("usuarioId")
+    @JoinColumn(name = "usuario_id", nullable = false)
+    private Usuario usuario;
+
+    @ManyToOne
+    @MapsId("conteudoId")
+    @JoinColumn(name = "conteudo_id", nullable = false)
+    private Conteudo conteudo;
+
+    @Column(nullable = false)
+    private LocalDateTime criadoEm;
+
+    @PrePersist
+    void prePersist() {
+        criadoEm = LocalDateTime.now();
+    }
 }
