@@ -2,41 +2,26 @@ package br.uniesp.si.techback.model;
 
 import jakarta.persistence.*;
 import lombok.*;
-import java.time.LocalDateTime;
+
 import java.util.UUID;
 
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(of = "id")
 @Entity
-@Table(name = "planos")
+@Table(name = "plano")
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor
+@EqualsAndHashCode(of = "id")
 public class Plano {
 
     @Id
     @GeneratedValue
     private UUID id;
 
-    @Column(nullable = false, unique = true, length = 100)
-    private String nome;
-
-    @Column(columnDefinition = "TEXT")
-    private String descricao;
-
-    @Column(name = "valor_mensal", nullable = false)
-    private Double valorMensal;
-
-    @Column(nullable = false, length = 10)
-    private String resolucao;
+    @Column(nullable = false, unique = true, length = 20)
+    private String codigo; // BASICO, PADRAO, PREMIUM
 
     @Column(nullable = false)
-    private Integer dispositivos;
+    private Integer limiteDiario;
 
-    @Column(name = "criado_em", nullable = false)
-    private LocalDateTime criadoEm;
-
-    @Column(name = "atualizado_em", nullable = false)
-    private LocalDateTime atualizadoEm;
+    @Column(nullable = false)
+    private Integer streamsSimultaneos;
 }
