@@ -1,50 +1,56 @@
 package br.uniesp.si.techback.model;
 
+import br.uniesp.si.techback.model.enums.PerfilUsuario;
 import jakarta.persistence.*;
 import lombok.*;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(of = "id")
 @Entity
 @Table(name = "usuarios")
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor
+@EqualsAndHashCode(of = "id")
 public class Usuario {
 
     @Id
     @GeneratedValue
     private UUID id;
 
-    @Column(name = "nome_completo", length = 150, nullable = false)
+    @Column(nullable = false, length = 150)
     private String nomeCompleto;
 
-    @Column(name = "data_nascimento", nullable = false)
+    @Column(nullable = false)
     private LocalDate dataNascimento;
 
-    @Column(nullable = false, length = 254, unique = true)
+    @Column(nullable = false, unique = true, length = 254)
     private String email;
 
-    @Column(name = "senha_hash", nullable = false, length = 60)
+    @Column(nullable = false, length = 60)
     private String senhaHash;
 
-    @Column(name = "cpf_cnpj", length = 14, unique = true)
+    @Column(unique = true, length = 14)
     private String cpfCnpj;
 
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
-    private String perfil;
+    private PerfilUsuario perfil;
 
-    @ManyToOne
-    @JoinColumn(name = "endereco_id")
-    private Endereco endereco;
-
-    @Column(name = "criado_em", nullable = false)
+    @Column(nullable = false)
     private LocalDateTime criadoEm;
 
-    @Column(name = "atualizado_em", nullable = false)
+    @Column(nullable = false)
     private LocalDateTime atualizadoEm;
+
+    @PrePersist
+    void prePersist() {
+        criadoEm = atualizadoEm = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    void preUpdate() {
+        atualizadoEm = LocalDateTime.now();
+    }
 }
