@@ -1,6 +1,7 @@
 package br.uniesp.si.techback.repository;
 
 import br.uniesp.si.techback.model.Assinatura;
+import br.uniesp.si.techback.model.enums.StatusAssinatura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -8,8 +9,7 @@ import java.util.UUID;
 
 public interface AssinaturaRepository extends JpaRepository<Assinatura, UUID> {
 
-    List<Assinatura> findByStatus(String status);
-
     List<Assinatura> findByUsuarioId(UUID usuarioId);
 
+    List<Assinatura> findByStatus(StatusAssinatura status);
 }
