@@ -1,6 +1,7 @@
 package br.uniesp.si.techback.repository;
 
 import br.uniesp.si.techback.model.Conteudo;
+import br.uniesp.si.techback.model.enums.TipoConteudo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -9,11 +10,14 @@ import java.util.UUID;
 
 public interface ConteudoRepository extends JpaRepository<Conteudo, UUID> {
 
-    List<Conteudo> findByGeneroIgnoreCase(String genero);
+    List<Conteudo> findAllByOrderByTituloAsc();
 
-    List<Conteudo> findByTipo(String tipo);
-
-    @Query("SELECT c FROM Conteudo c WHERE LOWER(c.titulo) LIKE LOWER(CONCAT('%', :q, '%'))")
-    List<Conteudo> buscarPorPalavraChave(String q);
-
+    @Query("""
+        SELECT c FROM Conteudo c
+        WHERE (:tipo IS NULL OR c.tipo = :tipo)
+          AND (:genero IS NULL OR LOWER(c.genero) = LOWER(:genero))
+          AND (:q IS NULL OR LOWER(c.titulo) LIKE LOWER(CONCAT('%', :q, '%')))
+        ORDER BY c.titulo
+    """)
+    List<Conteudo> filtrar(TipoConteudo tipo, String genero, String q);
 }
