@@ -5,48 +5,68 @@ import br.uniesp.si.techback.dto.plano.PlanoResponseDTO;
 import br.uniesp.si.techback.model.Plano;
 import br.uniesp.si.techback.repository.PlanoRepository;
 import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class PlanoService {
 
     private final PlanoRepository repository;
 
-    public PlanoService(PlanoRepository repository) {
-        this.repository = repository;
-    }
-
+    // ============================
+    // CRIAR
+    // ============================
     public PlanoResponseDTO criar(PlanoCreateDTO dto) {
-        Plano p = new Plano();
-        p.setNome(dto.nome());
-        p.setDescricao(dto.descricao());
-        p.setValorMensal(dto.valorMensal());
-        p.setResolucao(dto.resolucao());
-        p.setDispositivos(dto.dispositivos());
 
-        return toResponse(repository.save(p));
+        Plano plano = new Plano();
+        plano.setCodigo(dto.codigo());
+        plano.setLimiteDiario(dto.limiteDiario());
+        plano.setStreamsSimultaneos(dto.streamsSimultaneos());
+
+        return toResponse(repository.save(plano));
     }
 
+    // ============================
+    // BUSCAR POR ID
+    // ============================
     public PlanoResponseDTO buscar(UUID id) {
-        Plano p = repository.findById(id)
+        Plano plano = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Plano não encontrado"));
-        return toResponse(p);
+        return toResponse(plano);
     }
 
+    // ============================
+    // LISTAR
+    // ============================
     public List<PlanoResponseDTO> listar() {
-        return repository.findAll().stream().map(this::toResponse).toList();
+        return repository.findAll()
+                .stream()
+                .map(this::toResponse)
+                .toList();
     }
 
+    // ============================
+    // BUSCAR POR CÓDIGO
+    // ============================
+    public PlanoResponseDTO buscarPorCodigo(String codigo) {
+        Plano plano = repository.findByCodigo(codigo)
+                .orElseThrow(() -> new EntityNotFoundException("Plano não encontrado"));
+        return toResponse(plano);
+    }
+
+    // ============================
+    // MAPPER
+    // ============================
     private PlanoResponseDTO toResponse(Plano p) {
         return new PlanoResponseDTO(
                 p.getId(),
-                p.getNome(),
-                p.getValorMensal(),
-                p.getResolucao(),
-                p.getDispositivos()
+                p.getCodigo(),
+                p.getLimiteDiario(),
+                p.getStreamsSimultaneos()
         );
     }
 }
